requirements.txt
deepl>=1.18.0
anthropic>=0.25.0
typer>=0.12.0
rich>=13.0.0
python-dotenv>=1.0.0
